+++ components/icons/gatsby.tsx
@@ -1,9 +1,10 @@
 import React from 'react';
-import { ComponentProps, FC } from "react";
+import { ComponentProps } from 'react';
+
 export default function GatsbyIcon({
   width = 24,
   height = 24,
-}: ComponentProps<"svg">) {
+}: ComponentProps<'svg'>) {
   return (
     <svg
       width={width}
