--- lib/providers/query-provider.tsx
@@ -1,28 +1,0 @@
-'use client';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import { useState, ReactNode } from 'react';
-interface QueryProviderProps {
-  children: ReactNode;
-}
-export default function QueryProvider({ children }: QueryProviderProps) {
-  const [queryClient] = useState(
-    () =>
-      new QueryClient({
-        defaultOptions: {
-          queries: {
-            staleTime: 60 * 1000, 
-            gcTime: 5 * 60 * 1000, 
-            retry: 1,
-            refetchOnWindowFocus: false,
-          },
-        },
-      })
-  );
-  return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-      <ReactQueryDevtools initialIsOpen={false} />
-    </QueryClientProvider>
-  );
-}
