+++ package.json
@@ -22,8 +22,6 @@
     "@radix-ui/react-accordion": "^1.2.4",
     "@radix-ui/react-slot": "^1.2.0",
     "@tailwindcss/typography": "^0.5.16",
-    "@tanstack/react-query": "^5.77.0",
-    "@tanstack/react-query-devtools": "^5.77.0",
     "@vercel/analytics": "^1.5.0",
     "canvas": "^3.1.1",
     "class-variance-authority": "^0.7.1",
